# swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"Applying University Service API","description":"Applying service for University using Async "},"paths":{"/api/Application":{"get":{"tags":["Application"],"operationId":"GetAllAsyncApplication","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Application"}}}}},"post":{"tags":["Application"],"operationId":"Post","consumes":["application/json-patch+json","application/json","text/json","application/*+json","application/xml","text/xml","application/*+xml"],"produces":[],"parameters":[{"name":"value","in":"body","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Success"}}}},"/api/Application/{id}":{"get":{"tags":["Application"],"operationId":"findapplication","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}},"put":{"tags":["Application"],"operationId":"Put","consumes":["application/json-patch+json","application/json","text/json","application/*+json","application/xml","text/xml","application/*+xml"],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"},{"name":"value","in":"body","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Success"}}},"delete":{"tags":["Application"],"operationId":"Delete","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}}},"/api/Universities":{"get":{"tags":["Universities"],"operationId":"GetUniversity","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/University"}}}}}},"/api/Universities/{id}":{"get":{"tags":["Universities"],"operationId":"FindAsyncUniversity","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}}},"/api/Users":{"get":{"tags":["Users"],"operationId":"GetAllAsyncUser","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Users"}}}}},"post":{"tags":["Users"],"operationId":"PostUserAsync","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["text/plain","application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"users","in":"body","required":false,"schema":{"$ref":"#/definitions/Users"}}],"responses":{"201":{"description":"Success","schema":{"$ref":"#/definitions/Users"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/ProblemDetails"}}}}},"/api/Users/{id}":{"get":{"tags":["Users"],"operationId":"finduser","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}},"put":{"tags":["Users"],"operationId":"PutUserAsync","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["text/plain","application/json","text/json","application/xml","text/xml"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"},{"name":"users","in":"body","required":false,"schema":{"$ref":"#/definitions/Users"}}],"responses":{"204":{"description":"Success"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/ProblemDetails"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/ProblemDetails"}}}},"delete":{"tags":["Users"],"operationId":"DeleteAsync","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}}}},"definitions":{"Application":{"type":"object","properties":{"applicationId":{"format":"int32","type":"integer"},"userId":{"format":"int32","type":"integer"},"universityId":{"format":"int32","type":"integer"},"appliedDate":{"format":"date-time","type":"string"},"university":{"$ref":"#/definitions/University"},"user":{"$ref":"#/definitions/Users"}}},"University":{"type":"object","properties":{"universityId":{"format":"int32","type":"integer"},"universityName":{"type":"string"},"province":{"type":"string"},"city":{"type":"string"},"link":{"type":"string"},"latitude":{"type":"string"},"longitude":{"type":"string"},"application":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Application"}}}},"Users":{"type":"object","properties":{"userId":{"format":"int32","type":"integer"},"username":{"type":"string"},"password":{"type":"string"},"roll":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"application":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Application"}}}},"ProblemDetails":{"type":"object","properties":{"type":{"type":"string"},"title":{"type":"string"},"status":{"format":"int32","type":"integer"},"detail":{"type":"string"},"instance":{"type":"string"}},"additionalProperties":{"type":"object"}}}}